dfloat constructor to take long long instead of int64_t

+++ include/dfloat.h
@@ -121,7 +121,7 @@ namespace xu
     //  Conversions
     //  ===========
 
-    dfloat(int64_t value);
+    dfloat(long long value);
 
     dfloat(double value);
